+++ test/utils_t.py
@@ -134,39 +134,44 @@ class testUtils(unittest.TestCase):
     def test_parse_filter(self):
         """Test parse_filter function"""
         spec = {'file.size':1}
-        filter = 'monitor'
-        result = parse_filter(spec, filter)
-        expect = {filter: {'$exists':True}}
+        flt = 'monitor'
+        result = parse_filter(spec, flt)
+        expect = {flt: {'$exists':True}}
         self.assertEqual(expect, result)
 
-        filter = 'file.size'
-        result = parse_filter(spec, filter)
+        flt = 'file.size'
+        result = parse_filter(spec, flt)
         expect = {} # no filter must be added, since spec carry the condition
         self.assertEqual(expect, result)
 
-        filter = 'file.size=1'
-        result = parse_filter(spec, filter)
+        flt = 'file.size=1'
+        result = parse_filter(spec, flt)
         expect = {'file.size': 1}
         self.assertEqual(expect, result)
 
-        filter = 'file.size<1'
-        result = parse_filter(spec, filter)
+        flt = 'file.size<1'
+        result = parse_filter(spec, flt)
         expect = {'file.size': {'$lt': 1}}
         self.assertEqual(expect, result)
 
-        filter = 'file.size<=1'
-        result = parse_filter(spec, filter)
+        flt = 'file.size<=1'
+        result = parse_filter(spec, flt)
         expect = {'file.size': {'$lte': 1}}
         self.assertEqual(expect, result)
 
-        filter = 'file.size>1'
-        result = parse_filter(spec, filter)
+        flt = 'file.size>1'
+        result = parse_filter(spec, flt)
         expect = {'file.size': {'$gt': 1}}
         self.assertEqual(expect, result)
 
-        filter = 'file.size>=1'
-        result = parse_filter(spec, filter)
+        flt = 'file.size>=1'
+        result = parse_filter(spec, flt)
         expect = {'file.size': {'$gte': 1}}
+        self.assertEqual(expect, result)
+
+        flt = 'file.size!=1'
+        result = parse_filter(spec, flt)
+        expect = {'file.size': {'$ne': 1}}
         self.assertEqual(expect, result)
 
     def test_next_day(self):